Updates to term solving
We are almost at a point where we are done term solving.
Need to decide how to choose what terms we want to move onto the unifier with. Say I have
a term that covers 5 points, do I take that term? or do I just look for the first term which
satisfies a given point?